+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Change 'Foobu' tab title to 'Recommendations'
     - Limit birthday field maximum value to '2010-01-01'
     - Implement social media integration
+    - Add Terms and Conditions
 
 ## Version 0.8
     - Update lb-services.js